Revert "Merge pull request #52 from jupierce/bump_haproxy"
This reverts commit 663d9f2e8426dfe8c6410bf6ed3c54237471a29a, reversing
changes made to 32a3fb929461ff225eea39d5270356d91ab74d44.

haproxy-2.2.4 will not allow us to disable HTX anymore which means
client/servers will always get down-cased headers. As this may break
customers upgrading from OpenShift 4.6 => 4.7 we plan to add an API in
4.7 that will allow customers to specify case adjustment for various
routes. With that knob in place we will target 2.2.x for OpenShift 4.8
which is captured in this enhancement proposal:

  https://github.com/openshift/enhancements/pull/550

PR that reverts the router to haproxy20:

  https://github.com/openshift/router/pull/226