Add iptables shims to ipfailover-keepalived image
The registry.ci.openshift.org/ocp/4.8:base image that the
ipfailover-keepalived container is based on doesn't include iptables,
which keepalived needs. This patch uses the same technique that's used
in [openshift-sdn's](https://github.com/openshift/sdn) container images,
which is to utilize the host's iptables version.

With this commit, the ipfailover-keepalived image will require that the
host filesystem is mounted at `/host` inside the container, although
it's possible that just mounting the host's `/usr/sbin` at
`/host/usr/sbin/` will work as well